share window code behind between two projects, try to render to WPF D3DImage (unsuccessful)
