+++ VulkanInteropDirectX.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.33424.131
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Interop.WinUI3", "sample\Interop.WinUI3.csproj", "{6D31309D-22B0-4BB9-9DAE-2D4FFF8334B4}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Interop.Vulkan", "source\Interop.Vulkan.csproj", "{3757AAFE-C718-45BA-89F1-B180AB20FA0D}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "_solution", "_solution", "{0D6A86F1-0B28-4599-A312-DBF42932B130}"
@@ -14,22 +12,29 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "_solution", "_solution", "{
 		readme.md = readme.md
 	EndProjectSection
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Interop.WinUI3", "samples\winui\Interop.WinUI3.csproj", "{C3E46D51-9DF1-4826-B8EE-10CD7C882E0D}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Interop.WPF", "samples\wpf\Interop.WPF.csproj", "{39621384-C00E-4059-A332-77D74B3C4F24}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
 		Release|x64 = Release|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6D31309D-22B0-4BB9-9DAE-2D4FFF8334B4}.Debug|x64.ActiveCfg = Debug|x64
-		{6D31309D-22B0-4BB9-9DAE-2D4FFF8334B4}.Debug|x64.Build.0 = Debug|x64
-		{6D31309D-22B0-4BB9-9DAE-2D4FFF8334B4}.Debug|x64.Deploy.0 = Debug|x64
-		{6D31309D-22B0-4BB9-9DAE-2D4FFF8334B4}.Release|x64.ActiveCfg = Release|x64
-		{6D31309D-22B0-4BB9-9DAE-2D4FFF8334B4}.Release|x64.Build.0 = Release|x64
-		{6D31309D-22B0-4BB9-9DAE-2D4FFF8334B4}.Release|x64.Deploy.0 = Release|x64
 		{3757AAFE-C718-45BA-89F1-B180AB20FA0D}.Debug|x64.ActiveCfg = Debug|x64
 		{3757AAFE-C718-45BA-89F1-B180AB20FA0D}.Debug|x64.Build.0 = Debug|x64
 		{3757AAFE-C718-45BA-89F1-B180AB20FA0D}.Release|x64.ActiveCfg = Release|x64
 		{3757AAFE-C718-45BA-89F1-B180AB20FA0D}.Release|x64.Build.0 = Release|x64
+		{C3E46D51-9DF1-4826-B8EE-10CD7C882E0D}.Debug|x64.ActiveCfg = Debug|x64
+		{C3E46D51-9DF1-4826-B8EE-10CD7C882E0D}.Debug|x64.Build.0 = Debug|x64
+		{C3E46D51-9DF1-4826-B8EE-10CD7C882E0D}.Debug|x64.Deploy.0 = Debug|x64
+		{C3E46D51-9DF1-4826-B8EE-10CD7C882E0D}.Release|x64.ActiveCfg = Release|x64
+		{C3E46D51-9DF1-4826-B8EE-10CD7C882E0D}.Release|x64.Build.0 = Release|x64
+		{C3E46D51-9DF1-4826-B8EE-10CD7C882E0D}.Release|x64.Deploy.0 = Release|x64
+		{39621384-C00E-4059-A332-77D74B3C4F24}.Debug|x64.ActiveCfg = Debug|x64
+		{39621384-C00E-4059-A332-77D74B3C4F24}.Debug|x64.Build.0 = Debug|x64
+		{39621384-C00E-4059-A332-77D74B3C4F24}.Release|x64.ActiveCfg = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
