--- sample/App.xaml
@@ -1,16 +1,0 @@
-<Application
-  x:Class="Interop.WinUI3.App"
-  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-  <Application.Resources>
-    <ResourceDictionary>
-      <ResourceDictionary.MergedDictionaries>
-        <XamlControlsResources xmlns="using:Microsoft.UI.Xaml.Controls" />
-        <ResourceDictionary>
-          <SolidColorBrush x:Key="WindowCaptionBackground" Color="Transparent"/>
-          <SolidColorBrush x:Key="WindowCaptionBackgroundDisabled" Color="Transparent"/>
-        </ResourceDictionary>
-      </ResourceDictionary.MergedDictionaries>
-    </ResourceDictionary>
-  </Application.Resources>
-</Application>
