--- sample/MainWindow.xaml
@@ -1,38 +1,0 @@
-<Window
-  x:Class="Interop.WinUI3.MainWindow"
-  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-  xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-  mc:Ignorable="d"
-
-  Closed="OnWindowClosed">
-
-  <Window.SystemBackdrop>
-    <MicaBackdrop Kind="BaseAlt"/>
-  </Window.SystemBackdrop>
-
-  <Grid>
-    <Rectangle Name="titleBarRectangle" Height="25" Fill="Transparent"
-               HorizontalAlignment="Stretch" VerticalAlignment="Top"/>
-
-    <SwapChainPanel Margin="0" Name="swapchainPanel" Loaded="OnSwapchainPanelLoaded"/>
-
-    <Grid>
-      <Grid.RowDefinitions>
-        <RowDefinition Height="*"/>
-        <RowDefinition Height="2*"/>
-      </Grid.RowDefinitions>
-
-      <StackPanel Grid.Row="0" VerticalAlignment="Center" HorizontalAlignment="Center">
-        <Slider Minimum="1" Maximum="20" 
-                VerticalAlignment="Center" HorizontalAlignment="Stretch" ToolTipService.ToolTip="Rotation speed"
-                Value="{x:Bind vulkanInterop.RotationSpeed, Mode=TwoWay}"/>
-
-        <ToggleSwitch Margin="50,0,0,0" OffContent="Stopped" OnContent="Rotating"
-                      VerticalAlignment="Center" HorizontalAlignment="Center"
-                      Toggled="OnSwitchToggled"/>
-      </StackPanel>
-    </Grid>
-  </Grid>
-</Window>
