+++ samples/assets/DamagedHelmet.glb
