+++ samples/winui/assets/LockScreenLogo.scale-200.png
