+++ samples/winui/assets/SplashScreen.scale-200.png
