+++ samples/winui/assets/Square150x150Logo.scale-200.png
