+++ samples/winui/assets/Square44x44Logo.scale-200.png
