+++ samples/winui/assets/Square44x44Logo.targetsize-24_altform-unplated.png
