+++ samples/winui/assets/StoreLogo.png
