+++ samples/winui/assets/Wide310x150Logo.scale-200.png
