+++ samples/winui/properties/launchSettings.json
@@ -1,0 +1,8 @@
+{
+  "profiles": {
+    "Interop.WinUI3": {
+      "commandName": "MsixPackage",
+      "nativeDebugging": false
+    }
+  }
+}
