+++ samples/wpf/App.xaml
@@ -1,0 +1,9 @@
+﻿<Application 
+  x:Class="Interop.WPF.App"
+  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+  
+  StartupUri="MainWindow.xaml">
+
+  <Application.Resources/>
+</Application>
