+++ samples/wpf/App.xaml.cs
@@ -1,0 +1,6 @@
+﻿using System.Windows;
+
+namespace Interop.WPF
+{
+    public partial class App : Application { }
+}
