+++ samples/wpf/AssemblyInfo.cs
@@ -1,0 +1,7 @@
+using System.Windows;
+
+[assembly: ThemeInfo
+(
+    ResourceDictionaryLocation.None,
+    ResourceDictionaryLocation.SourceAssembly
+)]
