+++ source/Interop.Vulkan.csproj
@@ -17,8 +17,8 @@
   <ItemGroup>
     <PackageReference Include="SharpGLTF.Core" Version="1.0.0-alpha0030" />
 
-    <PackageReference Include="Silk.NET.Vulkan" Version="2.17.1" />
-    <PackageReference Include="Silk.NET.Vulkan.Extensions.KHR" Version="2.17.1" />
+    <PackageReference Include="Silk.NET.Vulkan" Version="2.18.0" />
+    <PackageReference Include="Silk.NET.Vulkan.Extensions.KHR" Version="2.18.0" />
   </ItemGroup>
 
   <ItemGroup>
